requirements.txt
flask
flask-cors
pyjwt==2.10.1
mysql-connector-python
